Refactor BaseKnexRepository into separate file
